graph: share the recycle item lookup between listing and restore, merge the recycle bin tests
